- fixed overlapping of error codes: internal reply code E_BAD_PROTO was   overlapping RFC3261 Temporarily Unavailable  (480) - dixed error code gnerated by Blacklist filtering. Instead of using the   generic send error, custom "473 Filtered destination" is used.
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1605 689a6050-402a-0410-94f2-e92a70836424

+++ forward.c
@@ -349,7 +349,7 @@ int forward_request( struct sip_msg* msg, struct proxy_l * p)
 			last_sock = send_sock;
 		}
 
-		if (check_blacklists(p->port,(unsigned short)p->proto,&to,buf,len)) {
+		if (check_blacklists( p->port, p->proto, &to, buf, len)) {
 			DBG("DEBUG:forward_request: blocked by blacklists\n");
 			ser_error=E_IP_BLOCKED;
 			continue;
